modules/devops/cicd-pipeline/functions/requirements.txt
google-cloud-deploy==1.13.0
google-cloud-pubsub==2.18.4
functions-framework==3.4.0